requirements.txt
alembic==1.2.1
aniso8601==8.0.0
attrs==19.2.0
bcrypt==3.1.7
cffi==1.12.3
click==7.0
coverage==4.5.4
enum-compat==0.0.2
eventlet==0.25.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
gem==0.1.12
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.3
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.9
Werkzeug==0.16.0
python-dotenv==0.10.3